refactor: extract the scripts' shared GitHub plumbing
`die`, `gh_token` and `detect_origin_repo` were byte-identical in
`export-github-item.py` and `pr-body.py` — 26 lines carrying one rule
each (the token-resolution order, the remote-URL shapes that count as
GitHub) that have to change together to stay correct.

They move to `scripts/lib/github.py`, imported as `from lib.github import
…`. A script run as `python3 scripts/<name>.py` puts `scripts/` on
`sys.path`, so the import resolves from any working directory — which
these scripts need, since they write relative to the caller's cwd while
living elsewhere.

Each script keeps its own `USER_AGENT` and docs root: those differ, or
are single-use, and pulling a one-line constant into a shared module
would cost a reader more than the duplication does.

This forks both files from the boilerplate they were vendored from, so
the same extraction is filed upstream to converge them again.
